fix: add an initialization phase to RedundantRBSSelector
The current implementation was transitioning unconditionally to
BOTH_SOURCES_VALID on init. Beyond causing a heisenbug (all 'start' should wait
for the event before going on), it is, you know, not true.

Add an initialization timeout during which the component will wait having enough
information to decide in which state it is, with a shortcut for
BOTH_SOURCES_VALID.